completion function in ChatCraftChat class
